bin/llm-repl-project/V3-minimal/pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "llm-repl-v3-minimal"
version = "0.1.0"
description = "LLM REPL V3 Minimal - Sacred Timeline with bottom-up approach"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "LLM REPL Team"},
]
dependencies = [
    "textual>=0.79.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "ollama>=0.3.0",
    "pyyaml>=6.0.2",
    "watchdog>=6.0.0",
    "tiktoken>=0.9.0",
    "textual-image>=0.8.3",
    "cairosvg>=2.8.2",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "types-PyYAML>=6.0.12.20250516",
    "pylint>=3.3.7",
    "radon>=6.0.1",
    "safety>=3.6.0",
    "bandit>=1.8.6",
    "vulture>=2.14",
    "pipdeptree>=2.27.0",
    "pip-licenses>=5.0.0",
]

[project.scripts]
llm-repl = "src.main:main"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
]
lint = [
    "mypy>=1.0.0",
    "ruff>=0.1.0", 
    "black>=23.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F841"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"